change price to int

+++ app/src/main/java/com/example/payeat/activities/MenuByTitleActivity.java
@@ -164,7 +164,7 @@ public class MenuByTitleActivity extends AppCompatActivity implements View.OnCli
 
             final String name=getItem(position).getName();
             final String desc=getItem(position).getDescription();
-            final double price =getItem(position).getPrice();
+            final int price =getItem(position).getPrice();
 
             TextView dishName = convertView.findViewById(R.id.dish_name_text);
             dishName.setText(name);
@@ -182,7 +182,7 @@ public class MenuByTitleActivity extends AppCompatActivity implements View.OnCli
                 bundle.putString("category", String_category);
                 bundle.putString("name", name);
                 bundle.putString("desc", desc);
-                bundle.putDouble("price", price);
+                bundle.putInt("price", price);
                 bundle.putBoolean("mode_manager", mode_manager);
                 if(mode_manager) {
                     bundle.putInt("tableNum", position);
